Adding code and notes

+++ Java Basics - Mosh/HelloWorld/src/Main.java
@@ -7,7 +7,7 @@ import java.util.Arrays;
 public class Main {
     public static void main(String[] args) {
         String x = "1.1";
-        int y = Integer.parseInt(x) + 2;
+        float y = Float.parseFloat(x) + 2;
         System.out.println(y);
     }
 }
